Moved all searchbars on front end to own component, adjusted all front end pages accordingly
