+++ lfucg-exactions/client/js/components/AgreementExisting.js
@@ -8,10 +8,10 @@ import Navbar from './Navbar';
 import Footer from './Footer';
 import Breadcrumbs from './Breadcrumbs';
 import Pagination from './Pagination';
+import SearchBar from './SearchBar';
 
 import {
     getPagination,
-    getAgreementQuery,
 } from '../actions/apiActions';
 
 import {
@@ -27,7 +27,6 @@ class AgreementExisting extends React.Component {
         const {
             currentUser,
             agreements,
-            onAgreementQuery,
         } = this.props;
 
         const agreements_list = agreements.length > 0 ? (
@@ -86,22 +85,7 @@ class AgreementExisting extends React.Component {
 
                 <Breadcrumbs route={this.props.route} />
 
-                <div className="row search-box">
-                    <form onChange={onAgreementQuery('query')} className="col-sm-10 col-sm-offset-1" >
-                        <fieldset>
-                            <div className="col-sm-2 col-xs-12">
-                                <label htmlFor="query" className="form-label">Search</label>
-                            </div>
-                            <div className="col-sm-10 col-xs-12">
-                                <input
-                                  type="text"
-                                  className="form-control"
-                                  placeholder="Search Agreements"
-                                />
-                            </div>
-                        </fieldset>
-                    </form>
-                </div>
+                <SearchBar />
 
                 <div className="inside-body">
                     <div className="container">
@@ -120,7 +104,6 @@ AgreementExisting.propTypes = {
     agreements: PropTypes.object,
     route: PropTypes.object,
     onComponentDidMount: PropTypes.func,
-    onAgreementQuery: PropTypes.func,
 };
 
 function mapStateToProps(state) {
@@ -134,16 +117,6 @@ function mapDispatchToProps(dispatch) {
     return {
         onComponentDidMount() {
             dispatch(getPagination('/agreement/'));
-        },
-        onAgreementQuery(field) {
-            return (e, ...args) => {
-                const value = typeof e.target.value !== 'undefined' ? e.target.value : args[1];
-                const update = {
-                    [field]: value,
-                };
-                dispatch(formUpdate(update));
-                dispatch(getAgreementQuery());
-            };
         },
     };
 }
