+++ lfucg-exactions/client/js/components/PaymentExisting.js
@@ -8,10 +8,10 @@ import Navbar from './Navbar';
 import Footer from './Footer';
 import Breadcrumbs from './Breadcrumbs';
 import Pagination from './Pagination';
+import SearchBar from './SearchBar';
 
 import {
     getPagination,
-    getPaymentQuery,
 } from '../actions/apiActions';
 
 import {
@@ -27,7 +27,6 @@ class PaymentExisting extends React.Component {
         const {
             currentUser,
             payments,
-            onPaymentQuery,
         } = this.props;
 
         const payments_list = payments.length > 0 &&
@@ -84,22 +83,7 @@ class PaymentExisting extends React.Component {
 
                 <Breadcrumbs route={this.props.route} />
 
-                <div className="row search-box">
-                    <form onChange={onPaymentQuery('query')} className="col-sm-10 col-sm-offset-1" >
-                        <fieldset>
-                            <div className="col-sm-2 col-xs-12">
-                                <label htmlFor="query" className="form-label">Search</label>
-                            </div>
-                            <div className="col-sm-10 col-xs-12">
-                                <input
-                                  type="text"
-                                  className="form-control"
-                                  placeholder="Search Payments"
-                                />
-                            </div>
-                        </fieldset>
-                    </form>
-                </div>
+                <SearchBar />
 
                 <div className="inside-body">
                     <div className="container">
@@ -118,7 +102,6 @@ PaymentExisting.propTypes = {
     payments: PropTypes.object,
     route: PropTypes.object,
     onComponentDidMount: PropTypes.func,
-    onPaymentQuery: PropTypes.func,
 };
 
 function mapStateToProps(state) {
@@ -132,16 +115,6 @@ function mapDispatchToProps(dispatch) {
     return {
         onComponentDidMount() {
             dispatch(getPagination('/payment/'));
-        },
-        onPaymentQuery(field) {
-            return (e, ...args) => {
-                const value = typeof e.target.value !== 'undefined' ? e.target.value : args[1];
-                const update = {
-                    [field]: value,
-                };
-                dispatch(formUpdate(update));
-                dispatch(getPaymentQuery());
-            };
         },
     };
 }
