+++ lfucg-exactions/client/js/components/PlatExisting.js
@@ -8,15 +8,11 @@ import Navbar from './Navbar';
 import Footer from './Footer';
 import Breadcrumbs from './Breadcrumbs';
 import Pagination from './Pagination';
+import SearchBar from './SearchBar';
 
 import {
     getPagination,
-    getPlatQuery,
 } from '../actions/apiActions';
-
-import {
-    formUpdate,
-} from '../actions/formActions';
 
 class PlatExisting extends React.Component {
     componentDidMount() {
@@ -27,7 +23,6 @@ class PlatExisting extends React.Component {
         const {
             currentUser,
             plats,
-            onPlatQuery,
         } = this.props;
 
         const plats_list = plats.length > 0 ? (
@@ -98,22 +93,7 @@ class PlatExisting extends React.Component {
 
                 <Breadcrumbs route={this.props.route} />
 
-                <div className="row search-box">
-                    <form onChange={onPlatQuery('query')} className="col-sm-10 col-sm-offset-1" >
-                        <fieldset>
-                            <div className="col-sm-2 col-xs-12">
-                                <label htmlFor="query" className="form-label">Search</label>
-                            </div>
-                            <div className="col-sm-10 col-xs-12">
-                                <input
-                                  type="text"
-                                  className="form-control"
-                                  placeholder="Search Plats"
-                                />
-                            </div>
-                        </fieldset>
-                    </form>
-                </div>
+                <SearchBar />
 
                 <div className="inside-body">
                     <div className="container">
@@ -132,7 +112,6 @@ PlatExisting.propTypes = {
     plats: PropTypes.object,
     route: PropTypes.object,
     onComponentDidMount: PropTypes.func,
-    onPlatQuery: PropTypes.func,
 };
 
 function mapStateToProps(state) {
@@ -146,16 +125,6 @@ function mapDispatchToProps(dispatch) {
     return {
         onComponentDidMount() {
             dispatch(getPagination('/plat/'));
-        },
-        onPlatQuery(field) {
-            return (e, ...args) => {
-                const value = typeof e.target.value !== 'undefined' ? e.target.value : args[1];
-                const update = {
-                    [field]: value,
-                };
-                dispatch(formUpdate(update));
-                dispatch(getPlatQuery());
-            };
         },
     };
 }
