+++ lfucg-exactions/client/js/components/ProjectExisting.js
@@ -8,10 +8,10 @@ import Navbar from './Navbar';
 import Footer from './Footer';
 import Breadcrumbs from './Breadcrumbs';
 import Pagination from './Pagination';
+import SearchBar from './SearchBar';
 
 import {
     getPagination,
-    getProjectQuery,
 } from '../actions/apiActions';
 
 import {
@@ -27,7 +27,6 @@ class ProjectExisting extends React.Component {
         const {
             currentUser,
             projects,
-            onProjectQuery,
         } = this.props;
 
         const projects_list = projects.length > 0 &&
@@ -86,22 +85,7 @@ class ProjectExisting extends React.Component {
 
                 <Breadcrumbs route={this.props.route} />
 
-                <div className="row search-box">
-                    <form onChange={onProjectQuery('query')} className="col-sm-10 col-sm-offset-1" >
-                        <fieldset>
-                            <div className="col-sm-2 col-xs-12">
-                                <label htmlFor="query" className="form-label">Search</label>
-                            </div>
-                            <div className="col-sm-10 col-xs-12">
-                                <input
-                                  type="text"
-                                  className="form-control"
-                                  placeholder="Search Projects"
-                                />
-                            </div>
-                        </fieldset>
-                    </form>
-                </div>
+                <SearchBar />
 
                 <div className="inside-body">
                     <div className="container">
@@ -120,7 +104,6 @@ ProjectExisting.propTypes = {
     projects: PropTypes.object,
     route: PropTypes.object,
     onComponentDidMount: PropTypes.func,
-    onProjectQuery: PropTypes.func,
 };
 
 function mapStateToProps(state) {
@@ -133,25 +116,7 @@ function mapStateToProps(state) {
 function mapDispatchToProps(dispatch) {
     return {
         onComponentDidMount() {
-            dispatch(getPagination('/project/'))
-            .then((data) => {
-                const account_update = {
-                    next: data.response.next,
-                    prev: data.response.prev,
-                    count: data.response.count,
-                };
-                dispatch(formUpdate(account_update));
-            });
-        },
-        onProjectQuery(field) {
-            return (e, ...args) => {
-                const value = typeof e.target.value !== 'undefined' ? e.target.value : args[1];
-                const update = {
-                    [field]: value,
-                };
-                dispatch(formUpdate(update));
-                dispatch(getProjectQuery());
-            };
+            dispatch(getPagination('/project/'));
         },
     };
 }
