+++ lfucg-exactions/client/js/components/SubdivisionExisting.js
@@ -8,10 +8,10 @@ import Navbar from './Navbar';
 import Footer from './Footer';
 import Breadcrumbs from './Breadcrumbs';
 import Pagination from './Pagination';
+import SearchBar from './SearchBar';
 
 import {
     getPagination,
-    getSubdivisionQuery,
 } from '../actions/apiActions';
 
 import {
@@ -28,7 +28,6 @@ class SubdivisionExisting extends React.Component {
         const {
             currentUser,
             subdivisions,
-            onSubdivisionQuery,
         } = this.props;
 
         const subdivisions_list = subdivisions && subdivisions.length > 0 &&
@@ -81,22 +80,7 @@ class SubdivisionExisting extends React.Component {
 
                 <Breadcrumbs route={this.props.route} />
 
-                <div className="row search-box">
-                    <form onChange={onSubdivisionQuery('query')} className="col-sm-10 col-sm-offset-1" >
-                        <fieldset>
-                            <div className="col-sm-2 col-xs-12">
-                                <label htmlFor="query" className="form-label">Search</label>
-                            </div>
-                            <div className="col-sm-10 col-xs-12">
-                                <input
-                                  type="text"
-                                  className="form-control"
-                                  placeholder="Search Subdivisions"
-                                />
-                            </div>
-                        </fieldset>
-                    </form>
-                </div>
+                <SearchBar />
 
                 <div className="inside-body">
                     <div className="container">
@@ -115,7 +99,6 @@ SubdivisionExisting.propTypes = {
     subdivisions: PropTypes.object,
     route: PropTypes.object,
     onComponentDidMount: PropTypes.func,
-    onSubdivisionQuery: PropTypes.func,
 };
 
 function mapStateToProps(state) {
@@ -129,16 +112,6 @@ function mapDispatchToProps(dispatch) {
     return {
         onComponentDidMount() {
             dispatch(getPagination('/subdivision/'));
-        },
-        onSubdivisionQuery(field) {
-            return (e, ...args) => {
-                const value = typeof e.target.value !== 'undefined' ? e.target.value : args[1];
-                const update = {
-                    [field]: value,
-                };
-                dispatch(formUpdate(update));
-                dispatch(getSubdivisionQuery());
-            };
         },
     };
 }
