+++ lfucg-exactions/client/js/constants/apiConstants.js
@@ -84,3 +84,4 @@ export const POST_ACCOUNT_LEDGER = 'POST_ACCOUNT_LEDGER';
 export const PUT_ACCOUNT_LEDGER = 'PUT_ACCOUNT_LEDGER';
 
 export const GET_PAGINATION = 'GET_PAGINATION';
+export const SEARCH_QUERY = 'SEARCH_QUERY';
