+++ lfucg-exactions/client/js/reducers/activeFormReducer.js
@@ -20,8 +20,8 @@ export default function activeFormReducer(state = {}, action) {
         return merge(state, action.update);
     case API_CALL_SUCCESS:
         if (action.endpoint === 'GET_PAGINATION' || action.endpoint.includes('QUERY')) {
-            const currentPage = window.location.hash.slice(1).toString() + '/';
-            return merge(state, { next: action.response.next, prev: action.response.prev, count: action.response.count, currentPage});
+            const currentPage = `${window.location.hash.slice(1)}/`;
+            return merge(state, { next: action.response.next, prev: action.response.prev, count: action.response.count, currentPage });
         }
         return state;
     case FORM_UPDATE:
